backend/requirements.txt
# API Framework
fastapi>=0.95.0
uvicorn>=0.21.1
python-multipart>=0.0.6
pydantic>=1.10.7

# PDF Processing
PyMuPDF>=1.21.1

# Entity Detection
presidio-analyzer>=2.2.33
presidio-anonymizer>=2.2.33
spacy>=3.5.2
transformers>=4.28.1
torch>=2.0.0
google-generativeai>=0.3.0

# GLiNER Integration
gliner_spacy>=0.0.7

# Utilities
python-dotenv>=1.0.0
requests>=2.28.2
pandas>=2.0.0
numpy>=1.24.3

gliner